requirements.txt
flask
werkzeug
SQLAlchemy
alembic
Flask-Migrate
Flask-Script
Flask-SqlAlchemy
Flask-Admin
Flask-Login
Flask-Security
psycopg2-binary
oauth2client
python-dotenv
flask-socketio
eventlet
flask-cors
factory-boy
bcrypt
